+++ CHANGELOG.md
@@ -10,6 +10,7 @@ omnibus-gitlab repository.
 - Rename geo_download_dispatch worker configuration
 - Rename geo_backfill worker configuration
 - Bump Git version to 2.13.0 b8a4bc4f
+- Upgrade PostgreSQL to 9.6.3 8f144d
 - Changing relative URL requires a hard reset ccd76ae2
 
 9.2.2